+++ example/lib/widgets/theme_selector.dart
@@ -13,13 +13,13 @@ class ThemeSelector extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     // Access the form from the parent context
-    final form = useFormContext<ProfileFields>(context);
+    final form = useFormContext(context);
 
-    return HookedFormField<ProfileFields, String>.explicit(
+    return HookedFormField<String, ProfileFields<String>>.explicit(
       form: form,
-      fieldHook: .theme,
+      fieldHook: ProfileFields.theme,
       notifyOnChange: true,
-      builder: (String? value, onChanged, error) {
+      builder: (value, onChanged, error) {
         return Column(
           crossAxisAlignment: .start,
           children: [