+++ lib/flutter_hook_form.dart
@@ -1,3 +1,5 @@
+export 'src/hooks/form_listen.dart';
+export 'src/hooks/use_field_value.dart';
 export 'src/hooks/use_form.dart';
 export 'src/hooks/use_form_context.dart';
 export 'src/messages/form_messages.dart';