+++ lib/src/hooks/use_form.dart
@@ -3,7 +3,6 @@ import 'package:flutter_hook_form/src/models/field_schema.dart';
 import 'package:flutter_hook_form/src/models/form_field_controller.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 
-
 /// A hook that provides a [FormFieldsController] to manage form field states.
 ///
 ///
@@ -13,16 +12,18 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 /// **Careful !**: this hook is a `flutter_hooks` hook and needs to be used
 /// inside a [HookWidget]. For more information about `flutter_hooks`, please
 /// refer to the [flutter_hooks documentation](https://pub.dev/packages/flutter_hooks).
-FormFieldsController<F> useForm<F extends FieldSchema>({
+FormFieldsController<F> useForm<F extends FieldSchema<dynamic>>({
   InitialFieldValues<F, dynamic>? initialValues,
   List<Object?> keys = const <Object>[],
 }) {
-  final controller = useMemoized(() {
+  final form = useMemoized(() {
     return FormFieldsController<F>(
       GlobalKey<FormState>(debugLabel: 'FormFieldsController'),
       initialValues: initialValues,
     );
   }, keys);
 
-  return useListenable(controller);
+  useEffect(() => form.dispose, [form]);
+
+  return form;
 }